+++ CHANGELOG.md
@@ -17,6 +17,7 @@ x.x.x Release notes (yyyy-MM-dd)
 * <Either mention core version or upgrade>
 * <Using Realm Core vX.Y.Z>
 * <Upgraded Realm Core from vX.Y.Z to vA.B.C>
+* Enabled `testNoMigrationOnSync`. ([#3312](https://github.com/realm/realm-js/issues/3312))
 
 10.19.2 Release notes (2022-6-20)
 =============================================================